done fixing the main part

+++ src/pages/MyItems.jsx
@@ -8,6 +8,7 @@ import { RiDeleteBin6Line } from "react-icons/ri";
 import { Link } from "react-router-dom";
 import UpdateItem from "./UpdateItem";
 import { useRef, useState } from "react";
+import Swal from "sweetalert2";
 
 const MyItems = () => {
   const { user } = useAuthHook();
@@ -57,7 +58,25 @@ const MyItems = () => {
   }
 
   const handleDelete = (id) => {
-    deleteItem(id);
+
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteItem(id);
+        // Swal.fire({
+        //   title: "Deleted!",
+        //   text: "Your file has been deleted.",
+        //   icon: "success"
+        // });
+      }
+    });
   };
 
   // console.log(items);
